requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-environ==0.11.2
django-multiselectfield==0.1.13
idna==3.10
paystackapi==2.1.3
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
